lib/common: relocate from usdpaa
Add common library from usdpaa repository into the newly created
repository usdpaa-apps.

Minor changes were made in code organization order to reflect the
usdpaa-apps repository structure:
- generic headers will be in 'include' directory
- application source code will be in 'src' directory
- generic libraries sources will be in 'lib' directory

There is a difference between usdpaa-apps build and the old usdpaa
build, namingly the application build rules do not have as a
dependency the driver libraries. This means that development on the
drivers will not be seen as a change in the application build process.

Signed-off-by: Bogdan-Constantin Popescu <bogdan.c.popescu@freescale.com>
Change-Id: I7bfa3da3d42fb344ba21393f8bc83c421bcc2db8
Reviewed-on: http://git.am.freescale.net:8181/31198
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>