ipsec_offload: Allow ICMP traffic in protocol=ANY IPSec outbound SA policies
Due to a design issue in the application PCD, the traffic that was
classified as ICMP was no longer checked for protocol=ANY policies. Because
of that, even though protocol=ANY was marked in the outbound SA policy,
the ICMP traffic was incorrectly bypassed.

This update chains the TCP/UDP table, which is designated as the placeholder
for the protocol=ANY policies, on the LOOKUP MISS relationship of the ICMP
table for both IPv4 and IPv6. This also requires some lookup key adjustments.

Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Change-Id: I1036bfac7973e7891808ba079485f8e77ea38a13
Reviewed-on: http://git.am.freescale.net:8181/32034
Reviewed-by: Bogdan Constantin Popescu <bogdan.c.popescu@freescale.com>
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>