offload apps: Fix incorrect use of old DPAA offloading library name
When all the offloading applications were migrated from the usdpaa
repository to the usdpaa-apps repository, the DPAA offloading user space
drivers were also separated into their own repository and software package.
When this was done, DPAA offloading user space library ceased to be
"libusdpaa_dpa_offload.a" and was renamed to "libdpa-offload.a". The
offloading applications though were not updated to link to this new
name. This update fixes this problem and links the offloading applications
with the correct "libdpa-offload.a" library.

Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Change-Id: I9c675f33f6947c6dd178d16c87734b02c6176a7b
Reviewed-on: http://git.am.freescale.net:8181/32033
Reviewed-by: Bogdan Constantin Popescu <bogdan.c.popescu@freescale.com>
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>