ipsec_offload: Process IPv6 route entries
The ipsec_offload application was failing to process the IPv6 route
entries due to lack of necessary support. This was causing it to miss
offloading the IPv6 routes that were not identical to the gateway.
This update adds the necessary support for the application to process
the IPv6 route messages from Netlink and properly offload all IPv6
post encrypt/decrypt routes.

Change-Id: I9b004c8fc86927fa9416bce921ff9da9b1670859
Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/32174
Reviewed-by: Radu-Andrei Bulie <Radu.Bulie@freescale.com>
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>