Makefile: fix parallel build
Because the driver libraries are not in the same repository, the
dependency of binary files on LDADD flags was dropped. This commit moves
all external libraries from LDADD to LDFLAGS (for all libraries from
LDADD flags the build system creates a build rule, but the libraries are
not in this repository anymore) and changes the linker flag order so
that the internal libraries which can depend on other external ones to
be correctly linked.

Now the applications may be compiled using multiple threads,
through -j flag.

Signed-off-by: Bogdan-Constantin Popescu <bogdan.c.popescu@freescale.com>
Change-Id: Id62ec899a4efdfa91bf3e3e52628d12ecbebda96
Reviewed-on: http://git.am.freescale.net:8181/33958
Reviewed-by: Marian-Cornel Chereji <marian.chereji@freescale.com>
Tested-by: Marian-Cornel Chereji <marian.chereji@freescale.com>
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>