dpa_offload: Fix implementation of inbound IP forwarding rules
The dpa_offload IPSec offloading demo application is showing an example
of how NF IP forwarding rules can be offloaded on the inbound direction.
The clear traffic on inbound is subject to these forwarding rules.
However, the application command line interface commands that were adding
and removing rules were not acquiring enough information from the user.

Also, the rule_get command was not found to be relevant to the
application's users and was removed.

Change-Id: Ife8dcfd8250603549c96bc2890ac4fd67cf98779
Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/35260
Reviewed-by: Radu-Andrei Bulie <Radu.Bulie@freescale.com>
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>