dpa_offload: Fix use of Netlink object marking for IP routes
The route cache change callback was always verifying if the netlink
route object is marked before doing anything. Marking is an easy way
to detect whether the route was already offloaded. If the object was
marked, the callback would abort. However, during route delete
operation it is normal for the object to be marked (i.e. to have been
previously offloaded). Hence bailing out on this condition doesn't
work in this case.

The use of Netlink object marking was updated and adapted for each
of the route ADD, CHANGE and DELETE operation.

Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Change-Id: I17afbb5d67a5b9b61120ccd949581439055520fc
Reviewed-on: http://git.am.freescale.net:8181/35263
Reviewed-by: Nicolae-Sebastian Grigore <sebastian.grigore@freescale.com>
Reviewed-by: Radu-Andrei Bulie <Radu.Bulie@freescale.com>
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>