simple_crypto: add swap option for block cipher descriptors
Descriptor creation functions for SEC processing have
an additional parameter which indicates if swapping is needed.
This is needed for platforms for which the endianness of the
host is not the same as the endianness of the SEC.
This patch does the necessary updates so that the descriptors
are generated properly for both little and big endian
platforms.

Change-Id: Ied7697dc8ba2b1f06a1344d00f49d0aaa60c0586
Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
(cherry picked from commit 07c6b11975635b5fb1728ff85835795d6ae981a6)