ipsec_offload: Add debug tools
Added a set of useful debug prints and CLI commands to facilitate
easy tracking of packet flows problems. The debug prints
and commands will only be available when the application is
compiled with the ENABLE_TRACE symbol. Also, in order to use
route table debug statistics commands the application PCD file
needs some modifications (namely, enabling statistics on the
routing tables). This change was not integrated however in order
not to impact overall performance.

Signed-off-by: Marian Chereji <marian.chereji@freescale.com>