libfslcrypto: fix comparison for results < 4B
Comparing of results with the reference test vectors
is done in 4B chunks. There is an issue when comparing
less that 4B on LE platforms. This patch converts the
remaining chunk so it can be compared in the same
manner.

Change-Id: I5c974f237c7a2a49125c9f530c4fc4fd15999200
Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
(cherry picked from commit 4e476ef469e00f3ab2e7044cae4acbffe4ed74d5)