dpa_offload: Avoid initializing offline ports Rx queues
This is a temporary WORKAROUND added to allow the application
to properly restart. Apparently, for some reason, if we create
and initialize the Rx default and Rx error queues for the 3
used offline ports, the application crashes in IP reassembly
infrastructure initialization when it is restarted. It appears
that simply by omitting the initialization and use of these
offline ports Rx queues allows the application to correctly
restart. As these queues are not currently used by the
application's packet flows they can be temporarily disabled
until the investigation for the real cause of this issue can
be completed.

Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Change-Id: I8fc23d4a586ed2e76b71630d6ffdca324d61968a
Reviewed-on: http://git.am.freescale.net:8181/35638
Reviewed-by: Radu-Andrei Bulie <Radu.Bulie@freescale.com>
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>