dpa_offload: Fix IP neighbour modification and removal
The Linux Netlink IP neighbour framework never actually issues
a NL_ACT_DEL message when users are typing "ip neigh del"
commands. It only changes that ARP entry's state to FAILED so
that the next time an application tries to send data to that
host, the Linux kernel triggers the L2 address refresh.

The neighbour state change callback function in the application
was adapted to consider any updates that takes an ARP entry to
the STALE or FAILED states as a "delete ARP entry" event.

Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Change-Id: I8faae703dd7f4d4759805cf22d1d45d369d88ea6
Reviewed-on: http://git.am.freescale.net:8181/35639
Reviewed-by: Radu-Andrei Bulie <Radu.Bulie@freescale.com>
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>