ipsec_offload: Filter out garbage PF_KEY socket messages
The PF_KEY socket which is used to access the IPSec SADB and
SPDB databases content can be spammed with XFRM broadcast
messages which are primarily intended for other applications.
The ipsec_offload application was only looking at the first
XFRM message received on its socket and, occasionally, it
was confused by garbage XFRM messages. Therefore, the SADB
and SPDB databases access mechanism was upgraded to use unique
identification information (based on PID, request sequence
number and message type) in its requests and to specifically
look for the answers to its requests in the pile of XFRM
messages received on the PF_KEY socket.

Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Change-Id: Ie66e3e7aa2db76a39adee44705958d67777a689a
Reviewed-on: http://git.am.freescale.net:8181/35786
Reviewed-by: Pinghua An <pinghua.an@freescale.com>
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>