dpa_offload: Avoid using Linux XFRM IPSec policy Ids
The NF layer supports a maximum range of 256 IPSec policy Ids for
each direction (inbound, outbound). The Linux XFRM appears to have
a different algorithm to generate the policy Ids which may or may
not be consecutive. It also provides unique policy Ids over the
entire IPSec engine which causes it to quickly overshoot 256 for
a fair number of tunnels.

Instead of just using the XFRM IPSec policy Ids for offloading,
the application was updated to generate its own, consecutive IPSec
policy Ids for each direction. This allows it to offload a lot more
policies than when just using the Linux XFRM policy Ids.

Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Change-Id: I4f69dcda0db6a112f845e6965f017e9463c410a1
Reviewed-on: http://git.am.freescale.net:8181/35864
Reviewed-by: Nicolae-Sebastian Grigore <sebastian.grigore@freescale.com>
Reviewed-by: Vakul Garg <vakul@freescale.com>