dpa_offload: Use inbound rule priority when deleting rules
The NF layer's IP forwarding module is using the source IP address,
destination IP address and ALSO the rule priority to identify a
forwarding rule. When deleting forwarding rules, these 3 pieces
of information are necessary to identify the inbound rule to delete.
The application was only requesting the IPSA and IPDA from the user
in ib_rule_del comands and it was always using priority=0 to search
for the rule to delete. This was causing the NF layer not to be able
to remove rules that were entered with a priority other than zero.

Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Change-Id: I54264bde99a4413379843f766e2e49bcada0fb93
Reviewed-on: http://git.am.freescale.net:8181/36044
Reviewed-by: Pinghua An <pinghua.an@freescale.com>
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>