ipsec_offload: Fix platform flag for VSP support
The platform flag that was used to enable Virtual Storage Profile
support was missing the T2080 platform. This caused problems on
T2080QDS and T2080RDB which had VSPs enabled in the DTS and PCD,
but not in the application.

For an easier maintenance, the platform flag was replaced with a
"VSP_SUPPORTED" flag which is set in the main application include
header depending on the current platform.

Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Change-Id: Ic861063d7daabe21103d00170e13d5fa778c08ba
Reviewed-on: http://git.am.freescale.net:8181/36191
Reviewed-by: Pinghua An <pinghua.an@freescale.com>
Reviewed-by: Vakul Garg <vakul@freescale.com>
Tested-by: Vakul Garg <vakul@freescale.com>