simple_crypto: fix SNOW_F8_F9 processing
LTE PDCP SNOW f8/f9 SEC descriptors are kept in a separate file
in simple_crypto (algo_desc.h). This file contains the SEC descriptors
in a hexadecimal format. Thus, the method used for all the other
descritptors (i.e. setting the "swap" parameter for the RTA
descriptor creation function). In order to accomodate running
on cores with a LE architecture and a BE DPAA subsystem,
the descriptors need to be converted to big endian[1].
The processing for LTE PDCP SNOW f8/f9 is being done by means
of inline job descriptors - a special type of descriptor which
is enqueued together with the frame to be processed and before
the actual frame. Of course, there is no shared descriptor
being used in this case. Because of this, the application needs
to calculate the dimension of the input buffer by adding to the
plaintext (or ciphertext for that matter) the length of the
inline job descriptor which performs the actual processing.
But the length calculation macro needs to take into account
the fact that the descriptor has been already 'swapped'. Thus,
a conversion from BE to the CPU endianess is needed for the
descriptor header word in order to properly extract the
length of the descriptor.

[1] The conversion is done word by word; this is not correct
    in the general case, since translating a descriptor to a
    different endianess is not done by merely byteswapping the
    words. In the particular case of the LTE PDCP SNOW f8/f9
    descriptor, it's sufficient to byteswap the words, since
    all the values to be loaded by the SEC HW are aligned to
    a 4 byte boundary.

Change-Id: Idb6e9e3b5c4fdadb9693c8ee2b1f12424a2e9d20
Signed-off-by: Alex Porosanu <alexandru.porosanu@freescale.com>
(cherry picked from commit 37dea9dba19f6b97dba7f7fb9bff5e0d03d6763f)