dpa_offload: Avoid creating multiple tables using the same Cc nodes
On the IPSec outbound direction the application was not recognizing
that sometimes the same Cc nodes were associated with multiple
protocols and it was creating different dpa_classifier tables. This
is a dangerous source of outbound policy tables inconsistency as
the same table would be modified from different interfaces w/o knowing
of each other.

The application was fixed to reuse existing IPSec outbound policy
tables for identical Cc nodes.

Signed-off-by: Marian Chereji <marian.chereji@freescale.com>