classifier_demo: Replace obsolete platform flags with FMan version flags
The dpa_offloading applications are performing different
configurations on different platforms according to the version
of the FMan that is present on that platform. In the past the
FMan version was somewhat guessed according to the platform flag
(B4860, P4080, etc.). Now, a specific FMan version flag is available
in fmlib and it is used especially for this purpose. There is no
further need to read the platform name. The use of build time
platform flags was therefore replaced with the use of the FMan
version flag.

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>