simple_proto: MBMS test vector fixes
The initial TVs were generated by using an initial value
of the LFSR set to all ones for both header as well as payload calculation.
According to the interoperability test conducted by Nokia/ALU, the initial
value of the LFSR needs to be all zeros. The changes done in RTA for the
descriptors change the output CRC values for both header and payload.
Thus, the TVs are updated by this patch for both negative as well
as positive tests.

Signed-off-by: Alex Porosanu <alexandru.porosanu@nxp.com>