Revert "build: Eliminate the use of implicit rules and rule-specific variables"
This is a cherry-picked fix from usdpaa-drivers. This Makefile is copied
straight from there and what follows is the commit message from that fix

The reverted patch removes builtin make variables that are essential to
proper functioning of the build process. In particular it removes the
variable "RM" which is actually used in clean target. By sheer luck, the RM
variable is not removed by this option in make < 4.0 but it is in recent
versions.
The consequence is that "clean" target fails with an obtuse message like
"make: ./include/fsl_sec/.Makefile.touch: Command not found" because
"rm" command is not substituted to the $(RM) variable which doesn't
exist.

This reverts commit 4c9064ff200592df718ed2de7bc50fd5b61b7f49 of
usdpaa-drivers.

Signed-off-by: Cristian Stoica <cristian.stoica@nxp.com>